+++ CHANGELOG.md
@@ -17,6 +17,8 @@
   * Show last four digits of vaulted cards, instead of "••• ••11", which truncates in larger fonts.
   * Remove support for Turkish localization
   * Remove `BraintreeUIKit` framework
+  * Use number pad for expiration date input on card form
+  * Require two-digit year for expiration date instead of four-digit year
 
 ## unreleased